Move middleware to Middleware namespace
